requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

# Linux Ubuntu Requirements - Wake Word Generator
# Optimized for Ubuntu 20.04+ with CUDA 11.8+ support

# Audio Processing
librosa==0.10.1
soundfile>=0.12.1
pydub>=0.25.1
scipy>=1.10.1

# Archive Handling
py7zr>=0.20.8
rarfile>=4.2
patool>=1.12.0  # Linux-compatible archive extraction

# Machine Learning - Linux CUDA optimized
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
tensorflow[and-cuda]==2.13.1  # Linux CUDA-enabled TensorFlow
keras==2.13.1
tensorboard==2.13.0
nvidia-ml-py>=12.535.77  # NVIDIA GPU monitoring
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0

# Visualization
matplotlib>=3.6.0
plotly>=5.0.0
seaborn>=0.12.0

# Web Interface
gradio==3.50.2  # Use newer Gradio for better Linux compatibility

# Text-to-Speech - Linux GPU optimized
piper-tts>=1.2.0
onnxruntime-gpu>=1.16.0  # Linux CUDA ONNX Runtime
elevenlabs>=1.0.0

# System utilities
tqdm>=4.64.0
pandas>=2.0.0
openwakeword>=0.5.0
python-dotenv>=1.0.0
psutil>=5.9.0  # System monitoring for Linux

# Audio system dependencies (Linux specific)
pyaudio>=0.2.11  # For microphone access
sounddevice>=0.4.6  # Alternative audio interface

# Development and debugging
ipython>=8.0.0  # Better debugging on Linux
jupyter>=1.0.0  # Notebook support for experimentation

# Additional Linux utilities
pyyaml>=6.0
requests>=2.28.0
click>=8.0.0  # CLI utilities
rich>=13.0.0  # Better console output